pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blockhost-broker"
version = "0.1.0"
description = "IPv6 tunnel broker for Blockhost installations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Blockhost Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    # On-chain dependencies
    "web3>=6.0.0",
    "eciespy>=0.4.0",
    "eth-account>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[project.scripts]
blockhost-broker = "blockhost_broker.main:main"
blockhost-broker-ctl = "blockhost_broker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
